+++ django_priority_batch/middleware.py
@@ -1,7 +1,11 @@
 """Django middleware for the prioritized batcher."""
+import logging
+
 from django.utils.deprecation import MiddlewareMixin
 
 from .prioritized_batcher import PrioritizedBatcher
+
+logger = logging.getLogger(__name__)
 
 
 class Middleware(MiddlewareMixin):
@@ -13,7 +17,18 @@ class Middleware(MiddlewareMixin):
 
     def process_request(self, request):
         """Process a request."""
-        PrioritizedBatcher.global_instance().start()
+        batcher = PrioritizedBatcher.global_instance()
+        if batcher.is_started:
+            # This can happen in old-style middleware if consequent middleware
+            # raises exception and thus `process_response` is not called.
+            # Described under 3rd point of differences:
+            # https://docs.djangoproject.com/en/1.11/topics/http/middleware/#upgrading-pre-django-1-10-style-middleware
+            batcher.rollback()
+            logger.warning(
+                "Uncommited batcher transaction from previous request was rollbacked."
+            )
+
+        batcher.start()
 
     def process_response(self, request, response):
         """Process a response."""