+++ docs/CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Fixed
+-----
+- Rollback batcher transaction at the beginning of middleware if is left
+  over from previous request
+
+
 ================
 1.1.0 2018-10-19
 ================